requirements.txt
pandas
yfinance
lxml
python-dotenv